% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{awncut.selection}
\alias{awncut.selection}
\title{This function outputs the selection of tuning parameters for the AWNCut method.}
\usage{
awncut.selection(X, Z, K, lambda, Tau, B = 500, L = 1000)
}
\arguments{
\item{X}{is an n x p1 matrix of n observations and p1 variables. The rows of
X into K clusters using the AWNCut method.}

\item{Z}{is an n x p2 matrix of n observations and p2 variables. Z is the assistant dataset.}

\item{K}{is the number of clusters.}

\item{lambda}{is a vector of tuning parameter lambda in the objective function.}

\item{Tau}{is a vector of tuning parameter tau in the objective function.}

\item{B}{is the number of iterations in the simulated annealing algorithm.}

\item{L}{is the temperature coefficient in the simulated annealing algorithm.}
}
\description{
This function outputs the selection of tuning parameters for the AWNCut method.
}
