% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AWNcut.UpdateCs}
\alias{AWNcut.UpdateCs}
\title{This function updates the clustering result of the AWNCut method.}
\usage{
AWNcut.UpdateCs(WX, WZ, K, Cs)
}
\arguments{
\item{WX}{is the weighed dataset X.}

\item{WZ}{is the weighed dataset Z.}

\item{K}{is the number of clusters.}

\item{Cs}{is the clustering result from the las iteration.}
}
\value{
Cs is the updated clustering result
}
\description{
This function updates the clustering result of the AWNCut method.
}
