% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AWNcut.UpdateWs}
\alias{AWNcut.UpdateWs}
\title{This function updates the feature selection result of the AWNCut method.}
\usage{
AWNcut.UpdateWs(X, Z, K, WX, WZ, b, Cs, ws, tau)
}
\arguments{
\item{X}{is an n x p1 matrix.}

\item{Z}{is an n x p2 matrix.}

\item{K}{is the number of clusters.}

\item{WX}{is the weighed dataset X.}

\item{WZ}{is the weighed dataset Z.}

\item{b}{is the iteration times.}

\item{Cs}{is the old clustering result.}

\item{ws}{is the old feature selection result.}

\item{tau}{is a vector of tuning parameter tau in the objective function.}
}
\value{
The output is a vector of the standardized updated weights.
}
\description{
This function updates the feature selection result of the AWNCut method.
}
