% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AWNcut.W}
\alias{AWNcut.W}
\title{This function calculates the weighed datasets for the AWNCut method.}
\usage{
AWNcut.W(X, Z, ws)
}
\arguments{
\item{X}{is a n x p1 matrix}

\item{Z}{is a n x p2 matrix}

\item{ws}{is a vector of weights for both X and Z datasets.}
}
\description{
This function calculates the weighed datasets for the AWNCut method.
}
