% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca.data.cna}
\alias{brca.data.cna}
\title{Data on copy number aberrations.}
\format{A data frame with n rows and p variables:
\describe{
  \item{variable1}{variable1 is something}
  \item{variable2}{variable2 is something else}
  ...
}}
\source{
\url{http:}
}
\usage{
brca.data.cna
}
\description{
A dataset containing CNAs from TCGA
}
\keyword{datasets}
