% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca.data.ge}
\alias{brca.data.ge}
\title{Data on gene expression from breast cancer patients.}
\format{A data frame with 873 patients and 334 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
brca.data.ge
}
\description{
A dataset containing gene expression measurements from TCGA.
}
\keyword{datasets}
