% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca.data.rppa}
\alias{brca.data.rppa}
\title{Data on protein measurements from breast cancer patients.}
\format{A data frame with 873 patients and 164 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
brca.data.rppa
}
\description{
A dataset containing protein measurements from TCGA.
}
\keyword{datasets}
