% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cesc.data.cna}
\alias{cesc.data.cna}
\title{Data on copy number aberrations from cervical cancer patients.}
\format{A data frame with 164 patients and 488 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
cesc.data.cna
}
\description{
A dataset containing copy number aberrations from TCGA.
}
\keyword{datasets}
