% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cesc.data.ge}
\alias{cesc.data.ge}
\title{Data on gene expression from breast cancer patients.}
\format{A data frame with 164 patients and 325 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
cesc.data.ge
}
\description{
A dataset containing gene expression measurements from TCGA.
}
\keyword{datasets}
