% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cesc.data.rppa}
\alias{cesc.data.rppa}
\title{Data on protein measurements from cervical cancer patients.}
\format{A data frame with 164 patients and 144 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
cesc.data.rppa
}
\description{
A dataset containing protein measurements from TCGA.
}
\keyword{datasets}
