% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DBI}
\alias{DBI}
\title{This function outputs the value of DBI for a clustering result.}
\usage{
DBI(X, K, Cs, ws)
}
\arguments{
\item{X}{is a n x p matrix with n observations and p variables.}

\item{K}{is the number of clusters.}

\item{Cs}{is a n x K matrix containing the clustering result of X. The entries
in Cs is either 1 or 0. If the ith observation is in the kth cluster, then Cs(i,k)=1,
otherwise, it equals to 0.}

\item{ws}{is a vector of the weights for the p variables. the length of ws equals p.}
}
\value{
The output is the value of DBI for a clustering result.
}
\description{
This function outputs the value of DBI for a clustering result.
}
