% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{distm}
\alias{distm}
\title{Given a data matrix X and matrix M of centroids calculate the distance
with respect to the centroids for each data matrix.}
\usage{
distm(M, X)
}
\arguments{
\item{M}{is a K x p matrix of cluster centroids.}

\item{X}{is a n x p matrix of p variables and n observations.}
}
\description{
This function will output a matrix K times p matrix of distances.
}
\details{
This is a function needed for the neokmeans algorithm.
}
