% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{ErrorRate}
\alias{ErrorRate}
\title{This function calculates the true error rate of a clustering result,
assuming that there are three clusters.}
\usage{
ErrorRate(X, n1, n2)
}
\arguments{
\item{X}{is a clustering result in matrix format.}

\item{n1}{is the size of the first cluster.}

\item{n2}{is the size of the second cluster.}
}
\value{
err is the true error rate of a clustering result.
}
\description{
This function calculates the true error rate of a clustering result,
assuming that there are three clusters.
}
