% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Kcs}
\alias{Kcs}
\title{This function transfers a clustering result into a matrix format from a vector.}
\usage{
Kcs(x)
}
\arguments{
\item{x}{is a clustering result in the vector format.}
}
\value{
A clustering result in the matrix format.
}
\description{
This function transfers a clustering result into a matrix format from a vector.
}
