% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{meanx}
\alias{meanx}
\title{Given a matrix U of K cluster membership and data X calculate the clusters' means.}
\usage{
meanx(U, X)
}
\arguments{
\item{X}{is a n x p matrix of p variables and n observations.}
}
\description{
This function will output K clusters' means.
}
\details{
Given a matrix U of K cluster membership calculated the clusters' means.
}
