% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{NCutY2R1}
\alias{NCutY2R1}
\title{Cluster the columns of Y into 2 groups with the help of external features in X.}
\usage{
NCutY2R1(Y, X, B = 3000, L = 1000, alpha = 0.5, nlambdas = 100,
  ncv = 5)
}
\arguments{
\item{Y}{is a n x p matrix of p variables and n observations.}

\item{X}{is a n x q matrix of q variables and n observations.}

\item{B}{is the number of iterations in the simulated annealing algorithm.}

\item{L}{is the temperature coefficient in the simulated annealing algorithm.}

\item{alpha}{is the coefficient of the elastic net penalty.}

\item{nlambdas}{is the number of tunning parameters in the elastic net.}

\item{ncv}{is the number of corss-validations in the elastic net.}
}
\value{
A list with the objective function, the clusters and the lambda penalty.
}
\description{
This function will output 2 groups of variables.
}
\details{
The algorithm minimizes a modified version of NCut through simulated annealing.
The clusers correspond to partitions that minimize this objective function.
The external information of X is incorporated by using ridge regression to predict Y.
}

