% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{NCutYLayer3R1}
\alias{NCutYLayer3R1}
\title{Cluster the columns of Z,Y and X into K channels.}
\usage{
NCutYLayer3R1(Z, Y, X, K = 2, B = 3000, L = 1000)
}
\arguments{
\item{Z}{is a n x p1 matrix of p1 variables and n observations.}

\item{Y}{is a n x p2 matrix of p2 variables and n observations.}

\item{X}{is a n x p3 matrix of p3 variables and n observations.}

\item{B}{is the number of iterations in the simulated annealing algorithm.}

\item{L}{is the temperature coefficient in the simulated annealing algorithm.}
}
\description{
This function will output K channels of variables.
}
\details{
The algorithm minimizes a modified version of NCut through simulated annealing.
The clusers correspond to partitions that minimize this objective function.
The external information of X is incorporated by using ridge regression to predict Y.
}

