% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{neoncut}
\alias{neoncut}
\title{Cluster the columns of X into K nonexhaustive overlapping clusters.}
\usage{
neoncut(X, K = 2, B = 3000, L = 1000, scale = F, mu_0 = 0.01,
  alpha = 0.5, beta = 0.5, dist = "gaussian", sigma = 1, MCMC = T)
}
\arguments{
\item{X}{is a n x p matrix of p variables and n observations.}

\item{B}{is the number of iterations in the simulated annealing algorithm.}

\item{L}{is the temperature coefficient in the simulated annealing algorithm.}
}
\description{
This function will output K clusters of variables.
}
\details{
The algorithm minimizes a modified version of NCut through simulated annealing.
The clusers correspond to partitions that minimize this objective function.
}
