% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Prob}
\alias{Prob}
\title{This function calculates the probability to accept the updated result
when updated value of the objectove function is smaller than the old one.
Note that argu2 must be greater than argu1.}
\usage{
Prob(argu1, argu2, L, b)
}
\arguments{
\item{argu1}{is the updates value of the objective function}

\item{argu2}{is the old value of the objective function}

\item{b}{is the iteration times.}
}
\description{
This function calculates the probability to accept the updated result
when updated value of the objectove function is smaller than the old one.
Note that argu2 must be greater than argu1.
}
