% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{spawn}
\alias{spawn}
\title{Cluster the columns of X into K clusters by giving a weight for each cluster while remaining sparse.}
\usage{
spawn(X, K = 2, B = 3000, L = 1000, N = 100, scale = T, lambda = 1,
  epsilon = 0, beta = 1, nstarts = 10, start = "default",
  dist = "gaussian", sigma = 1)
}
\arguments{
\item{X}{is a n x p matrix of p variables and n observations.}

\item{B}{is the number of iterations in the simulated annealing algorithm.}

\item{L}{is the temperature coefficient in the simulated annealing algorithm.}
}
\description{
This function will output K channels of variables.
}
\details{
The algorithm minimizes a modified version of NCut through simulated annealing.
The clusers correspond to partitions that minimize this objective function.
}
