% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Stability}
\alias{Stability}
\title{This function calculates the stablility of the simulation result}
\usage{
Stability(x)
}
\arguments{
\item{x}{is a 3 dimensional array.
The first dimension equals to sample size.
The second dimension equals to number og clusters.
The third dimension equals to the replication times.}
}
\value{
The output if the stability of the simulation result
}
\description{
This function calculates the stablility of the simulation result
}
