% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{swncut.cem}
\alias{swncut.cem}
\title{Cluster the columns of X into K clusters by giving a weight for each cluster while remaining sparse.}
\usage{
swncut.cem(X, K = 2, B = 30, N = 500, scale = T, q = 0.1,
  lambda = 1, dist = "correlation", epsilon = 0, beta = 1, sigma = 1)
}
\arguments{
\item{X}{is a n x p matrix of p variables and n observations.}

\item{K}{is the number of clusters.}

\item{B}{is the number of iterations in the cross entropy method.}

\item{N}{is the number of samples within each iteration of the cross entropy method.}

\item{scale}{equals TRUE if data X is to be scaled with mean 0 and variance 1.}

\item{lambda}{the tuning parameter of the penalty. Larger values shrink the weighted
cluster membership closer together (default = 1).}

\item{dist}{specifies the distance metric used for constructing the similarity matrix.
Options are 'gaussian', 'correlation' and 'euclidean' (default = 'gaussian').}

\item{epsilon}{values in the similarity matrix less than epsilon are set to 0 (default = 0).}

\item{beta}{when dist='correlation', beta is the exponent applied to each entry of the
similarity matrix.}

\item{sigma}{is the bandwith parameter when the dist metric chosen is 'gaussian' (default = 0.1).}

\item{nstarts}{the number of starting values also corresponding how many times simulated
annealing is run. Larger values provide better results but takes longer.}

\item{start}{if it equals 'default' then the starting value for all weights is 1/K. If
'random' then weights are sampled from a uniform distribution and then scaled to sum 1
per variable.}
}
\description{
This function will output K channels of variables.
}
\details{
The algorithm minimizes a modified version of NCut through simulated annealing.
The clusers correspond to partitions that minimize this objective function.
}
