% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AWNcut.OP}
\alias{AWNcut.OP}
\title{This function calculate the value of the objective function for the AWNCut method.}
\usage{
AWNcut.OP(X, Z, WX, WZ, Cs, tau)
}
\arguments{
\item{X}{is a n x p1 matrix.}

\item{Z}{is a n x p2 matrix.}

\item{WX}{is the weighed dataset X.}

\item{WZ}{is the weighed dataset Z.}

\item{Cs}{is clustering result.}

\item{tau}{is tuning parameter in the objective function.}
}
\description{
This function calculate the value of the objective function for the AWNCut method.
}
