% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca.data.cna}
\alias{brca.data.cna}
\title{Data on copy number aberrations from breast cancer patients.}
\format{A data frame with 873 patients and 515 gene names.}
\source{
\url{https://cancergenome.nih.gov/}
}
\usage{
brca.data.cna
}
\description{
A dataset containing copy number aberrations measurements from TCGA.
}
\keyword{datasets}
