% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ErrorRate}
\alias{ErrorRate}
\title{This function calculates the true error rate of a clustering result.}
\usage{
ErrorRate(X)
}
\arguments{
\item{X}{is a clustering result in matrix format.}
}
\value{
err is the true error rate of a clustering result.
}
\description{
This function calculates the true error rate of a clustering result.
}
