% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCutYX.R
\name{sncut}
\alias{sncut}
\title{Cluster the columns of Y into K groups using the NCut graph measure.}
\usage{
sncut(X, Z, K, lambda, B = 500, L = 1000)
}
\arguments{
\item{K}{is the number of clusters.}

\item{B}{is the number of iterations.}

\item{Y}{is a n x p matrix of p variables and n observations. The p columns of
Y will be clustered into K groups using NCut.}

\item{N}{is the number of samples per iterations.}

\item{scale}{equals TRUE if data Y is to be scaled with mean 0 and variance 1.}
}
\value{
A list with the final value of the objective function and
the clusters.
}
\description{
This function will output K clusters of the columns of Y.
}
\details{
The algorithm minimizes the NCut through the cross entropy method.
The clusters correspond to partitions that minimize this objective function.
}
